更新逻辑

+++ myblog-server/src/main/java/com/site/blog/model/entity/BlogCategory.java
@@ -19,7 +19,7 @@ import java.sql.Timestamp;
 public class BlogCategory  extends Model<BlogCategory> {
     @TableId(type = IdType.AUTO)
     private Integer relationId;
-    private Integer blogId;
+    private Long blogId;
     private Integer categoryId;
     private Timestamp createTime;
 }
